Algorithm

PSOJulia.sPSO — Method

sPSO(X_initial, V_initial, w, c1, c2, u1, u2, eta, sig, fobj, allN, up, lb)

Runs the smoothed Particle Swarm Optimisation algorithm on the list of inputs. The function returns a vector of recorded global best value estimates, and a matrix of their corresponding positions.

...

Arguments

X_initial: matrix of initial starting particle positions

V_initial: matrix of initial starting particle velocities

w: the intertia weight parameter

c1: the cognitive acceleration coefficient

c2: the social acceleration coefficient

u1: function for random variable generation for cognitive term

u2: function for random variable generation for social term

eta: step size parameter

sig: standard deviation parameter

fobj: objective function to be optimised

allN: list of iteration numbers to record results at

up: vector of upper bounds for each dimension

lb: vector of lower bounds for each dimension

...

Examples

allN = [10, 50, 100, 200, 400, Int(1e3), Int(3e3), Int(1e4)]
w = 1 - 0.729
c1 = 1.5
c2 = 1.5
en = 32
u1 = rand
u2 = rand
eta = 1
sig = 0
rep = 100
dim = 5

fobj = sphere
lb = fill(-100, dim)
up = fill(100, dim)
A = repeat(lb, 1, en)
B = repeat(up, 1, en)
X_initial = generate_random_matrix(A, B)
V_initial = generate_random_matrix(-(B-A), B-A)

values, positions = sPSO(X_initial, V_initial, w, c1, c2, u1, u2, eta, sig, fobj, allN, up, lb)

PSOJulia.PSO — Method

PSO(X_initial, V_initial, w, c1, c2, u1, u2, fobj, allN, up, lb)

Wrapper function that calls sPSO with eta=1 and sig=0. See sPSO.

PSOJulia.PSO

PSOJulia.sPSO