Algorithms

PSOJulia.CSO

PSOJulia.PSO

PSOJulia.sPSO

PSOJulia.CSO — Function

CSO(fobj, allN, separable, up, lb, modified)

Runs the Competitive Swarm Optimiser algorithm on the list of inputs. The function returns a vector of recorded global best value estimates, and a matrix of their corresponding positions.

Arguments

fobj: objective function to be optimised

allN: list of iteration numbers to record results at

separable: boolean of whether fobj is a separable function

up: scalar / vector of upper bounds for each dimension

lb: scalar / vector of lower bounds for each dimension

modified: optional argument that determines if two projections are done

sig: standard deviation parameter, optional if modified=false

Examples

julia> allN = [10, 50, 100, 200, 400, Int(1e3), Int(3e3), Int(1e4)];
julia> fobj = sphere;
julia> separable = true;
julia> up = fill(100., 5)
julia> lb = -up
julia> v = CSO(fobj, allN, separable, up, lb);
julia> v2 = CSO(fobj, allN, separable, up, lb, true);
julia> v3 = CSO(fobj, allN, separable, up, lb, true, sig=0.02);

PSOJulia.PSO — Function

PSO(X_initial, V_initial, w, c1, c2, u1, u2, fobj, allN, up, lb)

Wrapper function that calls sPSO with eta=1 and sig=0. See sPSO.

PSOJulia.sPSO — Function

sPSO(X_initial, V_initial, w, c1, c2, u1, u2, eta, sig, fobj, allN, up, lb, modified)

Runs the smoothed Particle Swarm Optimisation algorithm on the list of inputs. The function returns a vector of recorded global best value estimates, and a matrix of their corresponding positions.

Arguments

X_initial: matrix of initial starting particle positions

V_initial: matrix of initial starting particle velocities

w: the intertia weight parameter

c1: the cognitive acceleration coefficient

c2: the social acceleration coefficient

u1: function for random variable generation for cognitive term

u2: function for random variable generation for social term

eta: step size parameter

sig: standard deviation parameter

fobj: objective function to be optimised

allN: list of iteration numbers to record results at

up: scalar / vector of upper bounds for each dimension

lb: scalar / vector of lower bounds for each dimension

modified: optional argument that determines if two projections are done

Examples

julia> allN = [10, 50, 100, 200, 400, Int(1e3), Int(3e3), Int(1e4)];
julia> w = 1 - 0.729;
julia> c1 = 1.5;
julia> c2 = 1.5;
julia> en = 32;
julia> u1 = rand;
julia> u2 = rand;
julia> eta = 0.1;
julia> sig = 0.5;
julia> fobj = sphere;
julia> dim = 5;
julia> lb, up = -100.0, 100.0;
julia> A, B = matrix_bounds(lb, up, en, dim);
julia> v, p = sPSO(generate_particles(lb, up, en, dim), generate_velocities(lb, up, en, dim), 
                    w, c1, c2, u1, u2, eta, sig, fobj, allN, up, lb);
julia> v2, p2 = sPSO(generate_particles(lb, up, en, dim), generate_velocities(lb, up, en, dim), 
                    w, c1, c2, u1, u2, eta, sig, fobj, allN, up, lb, true);

References

R. Cheng and Y. Jin, "A Competitive Swarm Optimizer for Large Scale Optimization," in IEEE Transactions on Cybernetics, vol. 45, no. 2, pp. 191-204, Feb. 2015, doi: 10.1109/TCYB.2014.2322602.

J. Kennedy and R. Eberhart, "Particle swarm optimization," Proceedings of ICNN'95 - International Conference on Neural Networks, Perth, WA, Australia, 1995, pp. 1942-1948 vol.4, doi: 10.1109/ICNN.1995.488968.

Y. Shi and R. Eberhart, "A modified particle swarm optimizer," 1998 IEEE International Conference on Evolutionary Computation Proceedings. IEEE World Congress on Computational Intelligence (Cat. No.98TH8360), Anchorage, AK, USA, 1998, pp. 69-73, doi: 10.1109/ICEC.1998.699146.