Functions

Functions provided in the package are listed below. To use other objective functions, ensure that the new function takes in an argument of type Vector{Float64}, i.e. Array{Float64, 1}. The function should also return a single value of type Float64.

For example, if you were to add in a new function called sphere, the function can be written as follows:

function sphere(x::Vector{Float64})
    return sum(x.^2)
end

PSOJulia.sphere — Method

sphere(x)

Evaluates the sphere function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d x_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$.

Examples

julia> sphere([1;2;3;4])
30

PSOJulia.sumsquare — Method

sumsquare(x)

Evaluates the sumsquare function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d ix_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$.

Examples

julia> sumsquare([1;2;3;4])
100

PSOJulia.sphere

PSOJulia.sumsquare