Functions

Defining new functions

To use objective functions other than the ones provided in this package, ensure that the new function takes in an argument of type Vector{Float64}, i.e. Array{Float64, 1}. The function should also return a single value of type Float64.

For example, if you were to add in a new function called objective, the function can be written as follows:

function objective(x::Vector{Float64})
    return sum(x.^2)
end

Note that the new function should work on all sizes of Vector other than empty vectors unless otherwise stated. In this case, you should add in an assertion to ensure that the input will conform to the required size.

List of available functions

The function definitions and other information are taken with reference from Ab Wahab et. al[1] and Surjanovic and Bingham[2]. A total of 30 functions from the package are listed below.

PSOJulia.ackley

PSOJulia.beale

PSOJulia.bohachevsky1

PSOJulia.bohachevsky2

PSOJulia.bohachevsky3

PSOJulia.booth

PSOJulia.branin

PSOJulia.bukin6

PSOJulia.colville

PSOJulia.dixonprice

PSOJulia.dropwave

PSOJulia.sphere

PSOJulia.sumsquare

PSOJulia.ackley — Method

ackley(x)

Evaluates the ackley function on the vector x and returns the value.

\[f(x) = -a \exp \left(-b \sqrt{\frac{1}{d} \sum_{i=1}^d x_i^2} \right) - \exp \left(\frac{1}{d} cos(cx_i) + a + \exp(1) \right)\]

where $a = 0, b = 0.2, c = 2\pi$.

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-32 \leq x_i \leq 32$.

Examples

julia> ackley([1.0; 2.0; 3.0; 4.0])
8.434694444437465

PSOJulia.beale — Method

beale(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = (1.5 - x_1 + x_1 x_2)^2 + (2.25 - x_1 + x_1 x_2^2)^2 + (2.625 - x_1 + x_1 x_2^2)^2\]

The global minimum $f(x) = 0$ is attained at $x = (3, 0.5)$. The recommended bounds are $-4.5 \leq x_i \leq 4.5$.

Examples

julia> beale([1.0; 2.0])
126.935

PSOJulia.bohachevsky1 — Method

bohachevsky1(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i)-0.4\cos(4\pi x_{i+1})+0.7)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky1([1.5; 2.2; 3.6; 4.7])
102.78901699437496

PSOJulia.bohachevsky2 — Method

bohachevsky2(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i) \cos(4\pi x_{i+1})+0.3)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky2([1.5; 2.2; 3.6; 4.7])
100.56229490168752

PSOJulia.bohachevsky3 — Method

bohachevsky3(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i + 4\pi x_{i+1})+0.3)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky3([1.5; 2.2; 3.6; 4.7])
101.11363047737527

PSOJulia.booth — Method

booth(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = (x_1 + 2x_2 - 7)^2 + (2x_1 + x_2 - 5)^2\]

The global minimum $f(x) = 0$ is attained at $x = (1, 3)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> booth([1.0; 2.0])
5.0

PSOJulia.branin — Method

branin(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = a(x_2 - bx_1^2 + cx_1 - r)^2 + s(1 - t)\cos(x_1) + s\]

where $a = 1, b = \frac{5.1}{4\pi^2}, c = \frac{5}{\pi}, r = 6, s = 10, t = \frac{1}{8\pi}$.

The global minimum $f(x) = 0.397887$ is attained at $x = (9.42478, 2.475), (-\pi, 12.275), (\pi, 2.275)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> branin([1.0; 2.0])
21.62763539206238

PSOJulia.bukin6 — Method

bukin6(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = 100 \sqrt{\lvert x_2 - 0.01x_1^2 \rvert} + 0.01 \lvert x_1 + 10 \rvert\]

The global minimum $f(x) = 0$ is attained at $x = (-10, 1)$. The recommended bounds are $-15 \leq x_1 \leq -5, -3 \leq x_2 \leq 3$.

Examples

julia> bukin6([1.0; 2.0])
141.17735979665886

PSOJulia.colville — Method

colville(x)

Evaluates the colville function on the vector x of length 4 and returns the value.

\[f(x) = 100(x_1^2 - x_2)^2 + (x_1 - 1)^2 + (x_3 - 1)^2 + 90(x_3^2 - x_4)^2 + \\ 10.1((x_2 - 1)^2 + (x_4 - 1)^2) + 19.8(x_2 - 1)(x_4 - 1)\]

The global minimum $f(x) = 0$ is attained at $x = (1, 1, 1, 1)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> colville([1.0; 2.0; 3.0; 4.0])
2514.4

PSOJulia.dixonprice — Method

dixonprice(x)

Evaluates the dixonprice function on the vector x and returns the value.

\[f(x) = (x_1 - 1)^2 + \sum_{i=1}^n i(2x_i^2 - x_{i-1})^2\]

The global minimum $f(x) = 0$ is attained at $x_i = 2^{-\frac{2^i - 2}{2^i}}$. The recommended bounds are $-5 \leq x_i \leq 5$.

Examples

julia> dixonprice([1.0; 2.0; 3.0; 4.0])
4230.0

PSOJulia.dropwave — Method

dropwave(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = - \frac{1 + \cos\left( 12 \sqrt{x_1^2 + x_2^2} \right)}{0.5(x_1^2 + x_2^2) + 2}\]

The global minimum $f(x) = -1$ is attained at $x = (0, 0)$. The recommended bounds are $-5.12 \leq x_i \leq 5.12$.

Examples

julia> dropwave([1.0; 2.0])
-0.19357369461450374

PSOJulia.sphere — Method

sphere(x)

Evaluates the sphere function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d x_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> sphere([1.0; 2.0; 3.0; 4.0])
30

PSOJulia.sumsquare — Method

sumsquare(x)

Evaluates the sumsquare function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d ix_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-5.12 \leq x_i \leq 5.12$.

Examples

julia> sumsquare([1.0; 2.0; 3.0; 4.0])
100

1Ab Wahab MN, Nefti-Meziani S, Atyabi A (2015) A Comprehensive Review of Swarm Optimization Algorithms. PLoS ONE 10(5): e0122827. https://doi.org/10.1371/journal.pone.0122827

2Surjanovic, S. & Bingham, D. (2013). Virtual Library of Simulation Experiments: Test Functions and Datasets. Retrieved June 4, 2020, from http://www.sfu.ca/~ssurjano.