Functions

Defining new functions

To use objective functions other than the ones provided in this package, ensure that the new function takes in an argument of type Vector{Float64}, i.e. Array{Float64, 1}. The function should also return a single value of type Float64.

For example, if you were to add in a new function called objective, the function can be written as follows:

function objective(x::Vector{Float64})
    return sum(x.^2)
end

Note that the new function should work on all sizes of Vector other than empty vectors unless otherwise stated. In this case, you should add in an assertion to ensure that the input will conform to the required size.

List of available functions

The function definitions and other information are taken with reference from Ab Wahab et. al[1] and Surjanovic and Bingham[2]. A total of 30 functions from the package are listed below.

PSOJulia.ackley

PSOJulia.beale

PSOJulia.bohachevsky1

PSOJulia.bohachevsky2

PSOJulia.bohachevsky3

PSOJulia.booth

PSOJulia.branin

PSOJulia.bukin6

PSOJulia.colville

PSOJulia.dixonprice

PSOJulia.dropwave

PSOJulia.easom

PSOJulia.eggholder

PSOJulia.goldsteinprice

PSOJulia.griewank

PSOJulia.matyas

PSOJulia.mccormick

PSOJulia.michalewicz5

PSOJulia.perm

PSOJulia.powell

PSOJulia.rastrigin

PSOJulia.rosenbrock

PSOJulia.schaffer2

PSOJulia.schaffer4

PSOJulia.schwefel

PSOJulia.shubert

PSOJulia.sphere

PSOJulia.sumsquare

PSOJulia.trid6

PSOJulia.zakharov

PSOJulia.ackley — Method

ackley(x)

Evaluates the ackley function on the vector x and returns the value.

\[f(x) = -a \exp \left(-b \sqrt{\frac{1}{d} \sum_{i=1}^d x_i^2} \right) - \exp \left(\frac{1}{d} cos(cx_i) + a + \exp(1) \right)\]

where $a = 0, b = 0.2, c = 2\pi$.

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-32 \leq x_i \leq 32$.

Examples

julia> ackley([1.0; 2.0; 3.0; 4.0])
8.434694444437465

PSOJulia.beale — Method

beale(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = (1.5 - x_1 + x_1 x_2)^2 + (2.25 - x_1 + x_1 x_2^2)^2 + (2.625 - x_1 + x_1 x_2^2)^2\]

The global minimum $f(x) = 0$ is attained at $x = (3, 0.5)$. The recommended bounds are $-4.5 \leq x_i \leq 4.5$.

Examples

julia> beale([1.0; 2.0])
126.935

PSOJulia.bohachevsky1 — Method

bohachevsky1(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i)-0.4\cos(4\pi x_{i+1})+0.7)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky1([1.5; 2.2; 3.6; 4.7])
102.78901699437496

PSOJulia.bohachevsky2 — Method

bohachevsky2(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i) \cos(4\pi x_{i+1})+0.3)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky2([1.5; 2.2; 3.6; 4.7])
100.56229490168752

PSOJulia.bohachevsky3 — Method

bohachevsky3(x)

Evaluates the bohachevsky1 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} x_i^2 + 2x_{i+1}^2 -0.3\cos(3\pi x_i + 4\pi x_{i+1})+0.3)\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> bohachevsky3([1.5; 2.2; 3.6; 4.7])
101.11363047737527

PSOJulia.booth — Method

booth(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = (x_1 + 2x_2 - 7)^2 + (2x_1 + x_2 - 5)^2\]

The global minimum $f(x) = 0$ is attained at $x = (1, 3)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> booth([1.0; 2.0])
5.0

PSOJulia.branin — Method

branin(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = a(x_2 - bx_1^2 + cx_1 - r)^2 + s(1 - t)\cos(x_1) + s\]

where $a = 1, b = \frac{5.1}{4\pi^2}, c = \frac{5}{\pi}, r = 6, s = 10, t = \frac{1}{8\pi}$.

The global minimum $f(x) = 0.397887$ is attained at $x = (9.42478, 2.475), (-\pi, 12.275), (\pi, 2.275)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> branin([1.0; 2.0])
21.62763539206238

PSOJulia.bukin6 — Method

bukin6(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = 100 \sqrt{\lvert x_2 - 0.01x_1^2 \rvert} + 0.01 \lvert x_1 + 10 \rvert\]

The global minimum $f(x) = 0$ is attained at $x = (-10, 1)$. The recommended bounds are $-15 \leq x_1 \leq -5, -3 \leq x_2 \leq 3$.

Examples

julia> bukin6([1.0; 2.0])
141.17735979665886

PSOJulia.colville — Method

colville(x)

Evaluates the colville function on the vector x of length 4 and returns the value.

\[f(x) = 100(x_1^2 - x_2)^2 + (x_1 - 1)^2 + (x_3 - 1)^2 + 90(x_3^2 - x_4)^2 + \\ 10.1((x_2 - 1)^2 + (x_4 - 1)^2) + 19.8(x_2 - 1)(x_4 - 1)\]

The global minimum $f(x) = 0$ is attained at $x = (1, 1, 1, 1)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> colville([1.0; 2.0; 3.0; 4.0])
2514.4

PSOJulia.dixonprice — Method

dixonprice(x)

Evaluates the dixonprice function on the vector x and returns the value.

\[f(x) = (x_1 - 1)^2 + \sum_{i=1}^n i(2x_i^2 - x_{i-1})^2\]

The global minimum $f(x) = 0$ is attained at $x_i = 2^{-\frac{2^i - 2}{2^i}}$. The recommended bounds are $-5 \leq x_i \leq 5$.

Examples

julia> dixonprice([1.0; 2.0; 3.0; 4.0])
4230.0

PSOJulia.dropwave — Method

dropwave(x)

Evaluates the beale function on the vector x of length 2 and returns the value.

\[f(x) = - \frac{1 + \cos\left( 12 \sqrt{x_1^2 + x_2^2} \right)}{0.5(x_1^2 + x_2^2) + 2}\]

The global minimum $f(x) = -1$ is attained at $x = (0, 0)$. The recommended bounds are $-5.12 \leq x_i \leq 5.12$.

Examples

julia> dropwave([1.0; 2.0])
-0.19357369461450374

PSOJulia.easom — Method

easom(x)

Evaluates the easom function on the vector x of length 2 and returns the value.

\[f(x) = -\cos(x_1) \cos(x_2) \exp \left( -(x_1 - \pi)^2 - (x_2 - \pi)^2 \right)\]

The global minimum $f(x) = -1$ is attained at $x = (\pi, \pi)$. The recommended bounds are $-30 \leq x_i \leq 30$.

Examples

julia> easom([1.0; 2.0])
0.0006223571340136757

PSOJulia.eggholder — Method

eggholder(x)

Evaluates the eggholder function on the vector x of length 2 and returns the value.

\[f(x) = -(x_2 + 47) \sin \left( \sqrt{\left\lvert x_2 + \frac{x_1}{2} + 47 \right\rvert} \right) - x_1 \sin \left( \sqrt{\left\lvert x_1 - (x_2 + 47) \right\rvert} \right)\]

The global minimum $f(x) = -959.6407$ is attained at $x = (512, 404.2319)$. The recommended bounds are $-512 \leq x_i \leq 512$.

Examples

julia> eggholder([1.0; 2.0])
-34.08883356384574

PSOJulia.goldsteinprice — Method

goldsteinprice(x)

Evaluates the goldsteinprice function on the vector x of length 2 and returns the value.

\[f(x) = \left[1 + (x_1 + x_2 + 1)^2(19 - 14x_1 + 3x_1^2- 14x_2 + 6x_1x_2 + 3x_2^2)\right] \times \\ \left[30 + (2x_1 - 3x_2)^2(18 - 32x_1 + 12x_1^2 + 4x_2 - 36x_1x_2 + 27x_2^2)\right]\]

The global minimum $f(x) = 3$ is attained at $x = (0, -1)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> goldsteinprice([1.0; 2.0])
137150.0

PSOJulia.griewank — Method

griewank(x)

Evaluates the griewank function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d \frac{x_i^2}{4000} - \prod_{i=1}^d \cos \left( \frac{x_i}{\sqrt{i}} \right) + 1\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-600 \leq x_i \leq 600$.

Examples

julia> griewank([1.0; 2.0; 3.0; 4.0])
1.001870378003202

PSOJulia.matyas — Method

matyas(x)

Evaluates the matyas function on the vector x of length 2 and returns the value.

\[f(x) = 0.26(x_1^2 + x_2^2) - 0.48 x_1 x_2\]

The global minimum $f(x) = 0$ is attained at $x = (0, 0)$. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> matyas([1.0; 2.0])
0.3400000000000001

PSOJulia.mccormick — Method

mccormick(x)

Evaluates the mccormick function on the vector x of length 2 and returns the value.

\[f(x) = \sin(x_1 + x_2) + (x_1 - x_2)^2 - 1.5x_1 + 2.5x_2 + 1\]

The global minimum $f(x) = -1.9133$ is attained at $x = (-0.54719, -1.54719)$. The recommended bounds are $-1.5 \leq x_1 \leq 4, -3 \leq x_2 \leq 4$.

Examples

julia> mccormick([1.0; 2.0])
5.641120008059867

PSOJulia.michalewicz5 — Method

michalewicz5(x)

Evaluates the michalewicz5 function on the vector x of length 5 and returns the value.

\[f(x) = - \sum_{i=1}^5 \sin(x_i) \sin^{2m} \left( \frac{i x_i^2}{\pi} \right)\]

The global minimum $f(x) = -4.687658$ is attained at multiple global minima. The recommended bounds are $0 \leq x_i \leq \pi$.

Examples

julia> michalewicz5([1.0; 2.0; 3.0; 4.0; 5.0])
0.7958946159933006

PSOJulia.perm — Method

perm(x)

Evaluates the perm function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d \left( \sum_{i=1}^d \left( j^i + \beta \right) \left( \frac{x_j}{j} \right)^i - 1 \right)^2\]

where $\beta = 0.5$.

The global minimum $f(x) = 0$ is attained at $x = (1, 2, ... , d-1 , d)$. The recommended bounds are $-d \leq x_i \leq d$.

Examples

julia> perm([1.0; 2.0; 3.0; 4.0])
0.0

PSOJulia.powell — Method

powell(x)

Evaluates the powell function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d/4} \left[ \left( x_{4i-3} + 10x_{4i-2} \right)^2 + 5 \left( x_{4i-1} - x_{4i} \right) + \left( x_{4i-2} - 2x_{4i-1} \right)^4 + 10 \left( x_{4i-3} - x_{4i} \right) \right]\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> powell([1.0; 2.0; 3.0; 4.0])
1512.0

PSOJulia.rastrigin — Method

rastrigin(x)

Evaluates the rastrigin function on the vector x and returns the value.

\[f(x) = 10d + \sum_{i=1}^d \left[ x_i^2 - 10\cos(2\pi x_i) \right]\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-5.12 \leq x_i \leq 5.12$.

Examples

julia> rastrigin([1.0; 2.0; 3.0; 4.0])
30.0

PSOJulia.rosenbrock — Method

rosenbrock(x)

Evaluates the rosenbrock function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^{d-1} \left[ 100(x_{i+1} - x_i^2)^2 + (x_i - 1)^2 \right]\]

The global minimum $f(x) = 0$ is attained at $x = (1, ... , 1)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> rosenbrock([1.0; 2.0; 3.0; 4.0])
2705.0

PSOJulia.schaffer2 — Method

schaffer2(x)

Evaluates the schaffer2 function on the vector x of length 2 and returns the value.

\[f(x) = 0.5 + \frac{\sin^2(x_1^2 - x_2^2) - 0.5}{(1 + 0.001(x_1^2 + x_2^2))^2}\]

The global minimum $f(x) = 0$ is attained at $x = (0, 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> schaffer2([1.0; 2.0])
0.02467994027357423

PSOJulia.schaffer4 — Method

schaffer4(x)

Evaluates the schaffer4 function on the vector x of length 2 and returns the value.

\[f(x) = 0.5 + \frac{\cos^2(\sin(\lvert x_1^2 - x_2^2 \rvert)) - 0.5}{(1 + 0.001(x_1^2 + x_2^2))^2}\]

The global minimum $f(x) = 0.292579$ is attained at x = (0, 1.25313) or (0, -1.25313). The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> schaffer4([1.0; 2.0])
0.9754506010204307

PSOJulia.schwefel — Method

schwefel(x)

Evaluates the schwefel function on the vector x and returns the value.

\[f(x) = 418.9829d - \sum_{i=1}^d x_i \sin(\sqrt(\lvert x_i \rvert))\]

The global minimum $f(x) = -418.9829d$ is attained at $x = (420.9687, ... , 420.9687)$. The recommended bounds are $-500 \leq x_i \leq 500$.

Examples

julia> schwefel([1.0; 2.0; 3.0; 4.0])
-9.415272519067155

PSOJulia.shubert — Method

shubert(x)

Evaluates the shubert function on the vector x of length 2 and returns the value.

\[f(x) = \prod_{j=1}^2 \sum_{i=1}^5 i \cos((i + 1)x_j + i)\]

The global minimum $f(x) = -186.7309$ is attained at multiple global minima. The recommended bounds are $-10 \leq x_i \leq 10$.

Examples

julia> shubert([1.0; 2.0])
1.4675729549059044

PSOJulia.sphere — Method

sphere(x)

Evaluates the sphere function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d x_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-100 \leq x_i \leq 100$.

Examples

julia> sphere([1.0; 2.0; 3.0; 4.0])
30.0

PSOJulia.sumsquare — Method

sumsquare(x)

Evaluates the sumsquare function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d ix_i^2\]

The global minimum $f(x) = 0$ is attained at $x = (0, ... , 0)$. The recommended bounds are $-5.12 \leq x_i \leq 5.12$.

Examples

julia> sumsquare([1.0; 2.0; 3.0; 4.0])
100.0

PSOJulia.trid6 — Method

trid6(x)

Evaluates the trid6 function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d (x_i - 1)^2 - \sum_{i=2}^d x_i x_{i-1}\]

The global minimum $f(x) = -d(d+4)(d-1)/6$ is attained at $x_i = i(d + 1 - i)$. The recommended bounds are $-d^2 \leq x_i \leq d^2$.

Examples

julia> trid6([1.0; 2.0; 3.0; 4.0])
-6.0

PSOJulia.zakharov — Method

zakharov(x)

Evaluates the zakharov function on the vector x and returns the value.

\[f(x) = \sum_{i=1}^d x_i^2 - \left( \sum_{i=1}^d 0.5ix_i \right)^2 + \left( \sum_{i=1}^d 0.5ix_i \right)^4\]

The global minimum $f(x) = 0$ is attained at $x_i = (0 , ... , 0)$. The recommended bounds are $-5 \leq x_i \leq 10$.

Examples

julia> zakharov([1.0; 2.0; 3.0; 4.0])
50880.0

1Ab Wahab MN, Nefti-Meziani S, Atyabi A (2015) A Comprehensive Review of Swarm Optimization Algorithms. PLoS ONE 10(5): e0122827. https://doi.org/10.1371/journal.pone.0122827

2Surjanovic, S. & Bingham, D. (2013). Virtual Library of Simulation Experiments: Test Functions and Datasets. Retrieved June 4, 2020, from http://www.sfu.ca/~ssurjano.