Utilities

PSOJulia.generate_particles

PSOJulia.generate_particles

PSOJulia.generate_random_matrix

PSOJulia.generate_random_matrix

PSOJulia.generate_velocities

PSOJulia.generate_velocities

PSOJulia.matrix_bound

PSOJulia.matrix_bound

PSOJulia.generate_particles — Method

generate_particles(A::Matrix{Float64}, B::Matrix{Float64})

Wrapper function that calls generate_random_matrix(A, B) to generate initial particle positions.

PSOJulia.generate_particles — Method

generate_particles(A::Float64, B::Float64, en::Int64, dim::Int64)

Wrapper function that calls generate_random_matrix(A, B, en, dim) to generate initial particle positions.

PSOJulia.generate_random_matrix — Method

generate_random_matrix(A::Matrix{Float64}, B::Matrix{Float64})

Generates a matrix X of the same size as the input arguments A and B, each element containing a uniform random variable where each entry $X_{i,j}$ contains $U \sim (A_{i,j}, B_{i,j})$.

Examples

julia> A = [1 2; 3 4];
julia> B = [5 6; 7 8];
julia> generate_random_matrix(A, B)
2×2 Array{Float64,2}:
 4.1516   2.27863
 5.42611  6.95403

julia> C = [1 2 3; 4 5 6];
julia> generate_random_matrix(B, C)
ERROR: AssertionError: Unequal matrix sizes

PSOJulia.generate_random_matrix — Method

generate_random_matrix(A::Float64, B::Float64, en::Int64, dim::Int64)

Generates a matrix X of size (dim, en) each element containing a uniform random variable where each entry $X_{i,j}$ contains $U \sim (A, B)$.

Examples

julia> generate_random_matrix(-10.0, 10.0, 3, 2)
2×3 Array{Float64,2}:
 7.42334   0.601355  -0.882865
 6.26274  -4.92114    8.9034

PSOJulia.generate_velocities — Method

generate_velocities(A::Matrix{Float64}, B::Matrix{Float64})

Wrapper function that calls generate_random_matrix(-(B-A), B-A) to generate initial particle velocities.

PSOJulia.generate_velocities — Method

generate_velocities(A::Float64, B::Float64, en::Int64, dim::Int64)

Wrapper function that calls generate_random_matrix(-(B-A), B-A, en, dim) to generate initial particle velocities.

PSOJulia.matrix_bound — Method

matrix_bound(lb::Vector{Float64}, up::Vector{Float64}, en::Int64, dim::Int64)

Generates two matrices A and B of dimension (n, en), where n = dim is the length of lb and up. The columns of A are made up of copies of lb, and the columns of B are made up of copies of up.

Examples

julia> lb = [1.5; 2.5];
julia> up = [3.1; 4.7];
julia> A, B = matrix_bound(lb, up, 3, 2);
julia> A
2×3 Array{Float64,2}:
 1.5  1.5  1.5
 2.5  2.5  2.5
julia> B
2×3 Array{Float64,2}:
 3.1  3.1  3.1
 4.7  4.7  4.7

PSOJulia.matrix_bound — Method

matrix_bound(lb::Float64, up::Float64, en::Int64, dim::Int64)

Generates two matrices A and B of dimension (dim, en), where the entries of A contain lb and the entries of B contain up.

Examples

julia> A, B = matrix_bound(-5.12, 5.12, 2, 3);
julia> A
3×2 Array{Float64,2}:
 -5.12  -5.12
 -5.12  -5.12
 -5.12  -5.12
julia> B
3×2 Array{Float64,2}:
 5.12  5.12
 5.12  5.12
 5.12  5.12